% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_pkgload.R
\name{ab_pkgload}
\alias{ab_pkgload}
\title{ab_pkgload}
\usage{
ab_pkgload(pkgs, session = TRUE)
}
\arguments{
\item{pkgs}{A character vector containing the names of R package to be loaded}

\item{session}{A logical to indicate whether the session.info should be printed}
}
\value{

}
\description{
ab_pkgload
}
\examples{
ab_pkgload(pkgs = c("dplyr", "magrittr"))
}
