% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_roundnice.R
\name{ab_roundnice}
\alias{ab_roundnice}
\title{ab_roundnice}
\usage{
ab_roundnice(x, dig = 2, numeric = FALSE, less_than = FALSE)
}
\arguments{
\item{x}{a numeric value}

\item{dig}{the number of digits to be rounded to}

\item{numeric}{whether the result should be numeric or character}

\item{less_than}{whether values rounded to 0 should be reported with less than sign (eg, for P-values)}
}
\value{

}
\description{
ab_roundnice
}
\examples{
x <- c(0.4567, 0.0005001, 0.0004999, 0.0005000)
ab_roundnice(x = x, dig = 3, numeric = FALSE, less_than = TRUE)
ab_roundnice(x = x, dig = 3, numeric = FALSE, less_than = FALSE)
ab_roundnice(x = x, dig = 3, numeric = TRUE, less_than = FALSE)
ab_roundnice(x = x, dig = 3, numeric = TRUE, less_than = TRUE)
}
