% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize_first_letter.R
\name{capitalize_first_letter}
\alias{capitalize_first_letter}
\title{Capitalize first letter of string}
\usage{
capitalize_first_letter(phrase)
}
\arguments{
\item{phrase}{Vector of Character string. Any string of characters}
}
\value{
Vector of strings
}
\description{
Convert a string to lower case except first letter of the first word.
}
