% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regseq.R
\name{create_reg_seq}
\alias{create_reg_seq}
\title{Create regular sequence of fixed length}
\usage{
create_reg_seq(seq, fill = "0", prefix = "", suffix = "")
}
\arguments{
\item{seq}{The sequence to complete.}

\item{fill}{The padding character.}

\item{prefix}{Prefix string.}

\item{suffix}{Suffix string..}
}
\value{
A sequence of
}
\description{
Create regular sequence of fixed length
}
\examples{
create_reg_seq(c(8:12, 98:102), prefix = "file", suffix = ".txt")
}
