% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{fetch_installed_packages}
\alias{fetch_installed_packages}
\title{Fetch installed packages}
\usage{
fetch_installed_packages(all_info = FALSE)
}
\arguments{
\item{all_info}{Logical.}
}
\value{
A tibble.
}
\description{
Fetch locally installed packages with informations about the sources.
}
\examples{
fetch_installed_packages()
}
