% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_textstring.R
\name{find_textstring}
\alias{find_textstring}
\title{Find snippets of text human readable files}
\usage{
find_textstring(textSnippet, folder, fileType)
}
\arguments{
\item{textSnippet}{Character string. A snippet of text}

\item{folder}{Character string. Folder to search. (Full path required)}

\item{fileType}{Character String. File type extension. eg ("R","Rmd","txt")}
}
\value{
Character vector. Names of the files that contain the text
}
\description{
Scans all files in specified folder for \code{textSnippet}
}
\examples{
\dontrun{
#search all R files in the project root for the string "parameter"
find_text_string("parameter", here::here(),"R")

# search all Rmd files in the vignettes folder for the string "help"
find_text_string("help", here::here("vignettes"),"Rmd")

}

}
