% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{package_source}
\alias{package_source}
\title{Find package source}
\usage{
package_source(pkg)
}
\arguments{
\item{pkg}{Name of the package}
}
\value{
A 1*2 tibble
}
\description{
Find package source
}
\examples{
abutils:::package_source("stats")
abutils:::package_source("dplyr")
}
